+++ src/assets/ba_data/python/baclassic/_appsubsystem.py
@@ -819,6 +819,8 @@ class ClassicAppSubsystem(babase.AppSubsystem):
 
     def invoke_main_menu_ui(self) -> None:
         """Bring up main menu ui."""
+        print('INVOKING MAIN MENU UI')
+
         # Bring up the last place we were, or start at the main menu otherwise.
         app = bauiv1.app
         env = app.env
@@ -830,18 +832,15 @@ class ClassicAppSubsystem(babase.AppSubsystem):
                 # UI stuff fails now in headless builds; avoid it.
                 pass
             else:
-                # main_menu_location = (
-                #     bascenev1.app.ui_v1.get_main_menu_location()
-                # )
 
-                # When coming back from a kiosk-mode game, jump to
-                # the kiosk start screen.
+                # When coming back from a kiosk-mode game, jump to the
+                # kiosk start screen.
                 if env.demo or env.arcade:
                     # pylint: disable=cyclic-import
                     from bauiv1lib.kiosk import KioskWindow
 
                     app.ui_v1.set_main_window(
-                        KioskWindow(), from_window=False  # Disable check here.
+                        KioskWindow(), is_top_level=True, suppress_warning=True
                     )
                 # ..or in normal cases go back to the main menu
                 else:
@@ -912,17 +911,3 @@ class ClassicAppSubsystem(babase.AppSubsystem):
                         is_top_level=True,
                         suppress_warning=True,
                     )
-
-                # attempt to show any pending offers immediately.
-                # If that doesn't work, try again in a few seconds
-                # (we may not have heard back from the server)
-                # ..if that doesn't work they'll just have to wait
-                # until the next opportunity.
-                # if not specialoffer.show_offer():
-
-                #     def try_again() -> None:
-                #         if not specialoffer.show_offer():
-                #             # Try one last time..
-                #             bauiv1.apptimer(2.0, specialoffer.show_offer)
-
-                #     bauiv1.apptimer(2.0, try_again)