+++ src/ballistica/base/graphics/component/render_component.h
@@ -5,7 +5,12 @@
 
 #include <vector>
 
-#include "ballistica/base/graphics/renderer/renderer.h"
+#include "ballistica/base/assets/mesh_asset.h"
+#include "ballistica/base/graphics/graphics.h"
+#include "ballistica/base/graphics/renderer/render_pass.h"
+#include "ballistica/base/graphics/support/frame_def.h"
+#include "ballistica/base/graphics/support/render_command_buffer.h"
+#include "ballistica/shared/math/rect.h"
 
 namespace ballistica::base {