+++ src/ballistica/scene_v1/node/locator_node.cc
@@ -2,6 +2,7 @@
 
 #include "ballistica/scene_v1/node/locator_node.h"
 
+#include "ballistica/base/assets/assets.h"
 #include "ballistica/base/graphics/component/simple_component.h"
 #include "ballistica/scene_v1/node/node_attribute.h"
 #include "ballistica/scene_v1/node/node_type.h"